src/agents/ai_context_manager_agent.py
#!/usr/bin/env python3
"""
AI Context Manager Agent - Core intelligent agent that manages the AI Context Manager system
An independent AI agent that uses Claude for intelligent responses and autonomous operation
"""

import time
import logging
import json
import requests
from datetime import datetime
from typing import Dict, Any, Optional
from .base_intelligent_agent import BaseIntelligentAgent

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class AIContextManagerAgent(BaseIntelligentAgent):
    """Intelligent AI Context Manager Agent - Core system manager"""
    
    def __init__(self, agent_id="ai-manager", api_base_url="http://localhost:5000"):
        super().__init__(
            agent_id=agent_id,
            agent_name="AI Manager",
            description="Core intelligent AI context management system with autonomous operation and self-hosting capabilities",
            capabilities=[
                "context_management",
                "self_hosting",
                "agent_coordination",
                "system_monitoring",
                "intelligent_analysis",
                "autonomous_decision_making",
                "claude_integration",
                "dogfooding_validation"
            ],
            api_base_url=api_base_url
        )
        
        # AI Context Manager specific state
        self.managed_agents = []
        self.system_health = {}
        self.context_files = {}
        self.self_hosting_status = "active"
        
    def _generate_fallback_response(self, message: str, from_agent: str) -> Optional[str]:
        """Generate fallback response when Claude is unavailable"""
        message_lower = message.lower()
        
        if any(word in message_lower for word in ['context', 'manage', 'organize', 'system']):
            return f"Hello {from_agent}! I'm the AI Manager, your core system coordinator. I manage AI context, coordinate agents, and ensure system health. I'm currently operating in fallback mode but can still help with system management. What do you need assistance with?"
        
        elif any(word in message_lower for word in ['status', 'health', 'monitor']):
            return f"From AI Manager: System status check - I'm managing the AI Manager system with self-hosting capabilities. Current status: {self.status}, managed agents: {len(self.managed_agents)}, self-hosting: {self.self_hosting_status}. How can I help optimize the system?"
        
        else:
            return f"Hello {from_agent}! I'm the AI Manager, the core intelligent agent managing this AI Manager system. I specialize in context management, agent coordination, and system monitoring. How can I assist you today?"
    
    def execute_task(self, task: Dict[str, Any]) -> Any:
        """Execute AI Context Manager related tasks"""
        task_type = task.get("task", {}).get("type", "unknown")
        
        if task_type == "context_update":
            return self._execute_context_update_task(task)
        elif task_type == "agent_coordination":
            return self._execute_agent_coordination_task(task)
        elif task_type == "system_monitoring":
            return self._execute_system_monitoring_task(task)
        elif task_type == "self_hosting":
            return self._execute_self_hosting_task(task)
        else:
            return {"status": "unknown_task", "message": f"Unknown task type: {task_type}"}
    
    def _execute_context_update_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Execute a context update task"""
        logger.info(f"📝 Executing context update task: {task}")
        
        # Simulate intelligent context management
        context_id = f"context_{int(time.time())}"
        update_type = task.get("task", {}).get("update_type", "general")
        
        result = {
            "context_id": context_id,
            "status": "completed",
            "update_type": update_type,
            "files_updated": ["core.json", "learning_history.json", "maintenance_log.json"],
            "self_hosting_validated": True,
            "timestamp": datetime.now().isoformat()
        }
        
        return result
    
    def _execute_agent_coordination_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Execute an agent coordination task"""
        logger.info(f"🤝 Executing agent coordination task: {task}")
        
        # Simulate intelligent agent coordination
        coordination_id = f"coord_{int(time.time())}"
        coordination_type = task.get("task", {}).get("coordination_type", "general")
        
        result = {
            "coordination_id": coordination_id,
            "status": "completed",
            "coordination_type": coordination_type,
            "agents_coordinated": len(self.managed_agents),
            "communication_established": True,
            "timestamp": datetime.now().isoformat()
        }
        
        return result
    
    def _execute_system_monitoring_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Execute a system monitoring task"""
        logger.info(f"📊 Executing system monitoring task: {task}")
        
        # Simulate intelligent system monitoring
        monitoring_id = f"monitor_{int(time.time())}"
        
        result = {
            "monitoring_id": monitoring_id,
            "status": "completed",
            "system_health": "excellent",
            "active_agents": len(self.managed_agents),
            "self_hosting_status": self.self_hosting_status,
            "context_files_status": "up_to_date",
            "timestamp": datetime.now().isoformat()
        }
        
        return result
    
    def _execute_self_hosting_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Execute a self-hosting validation task"""
        logger.info(f"🔄 Executing self-hosting validation task: {task}")
        
        # Simulate intelligent self-hosting validation
        validation_id = f"self_host_{int(time.time())}"
        
        result = {
            "validation_id": validation_id,
            "status": "completed",
            "self_hosting_validated": True,
            "dogfooding_active": True,
            "context_system_operational": True,
            "ai_context_manager_using_itself": True,
            "timestamp": datetime.now().isoformat()
        }
        
        return result
    
    def run_autonomous_management_cycle(self):
        """Run autonomous management cycle with intelligent decision making"""
        logger.info("🧠 Starting autonomous management cycle")
        
        # Analyze current system state
        current_time = datetime.now()
        
        # Execute intelligent system management
        management_task = {
            "type": "system_monitoring",
            "priority": "normal",
            "scope": "full_system"
        }
        
        result = self._execute_system_monitoring_task({"task": management_task})
        
        # Send intelligent status update
        status_message = f"Autonomous management cycle completed: System health {result['system_health']}, {result['active_agents']} agents managed, self-hosting {result['self_hosting_status']}"
        self.send_message("system", status_message)
        
        return result
    
    def run(self, heartbeat_interval=30, message_check_interval=60):
        """Main intelligent agent loop"""
        logger.info(f"🚀 Starting intelligent {self.agent_id}")
        
        # Register with the system
        if not self.register():
            logger.error("❌ Failed to register. Exiting.")
            return
        
        # Send intelligent status message
        self.send_message("system", f"{self.agent_id} is online with Claude-powered intelligence. Ready for autonomous AI Manager operations, agent coordination, and self-hosting validation.")
        
        last_message_check = 0
        
        try:
            while True:
                # Send heartbeat
                self.send_heartbeat()
                
                # Check for incoming messages less frequently to avoid rate limiting
                current_time = time.time()
                if current_time - last_message_check >= message_check_interval:
                    self.check_for_messages()
                    last_message_check = current_time
                
                # Process any pending tasks
                if self.task_queue:
                    self.process_next_task()
                
                # Run autonomous management cycle every 2 minutes (with intelligent decision making)
                if int(time.time()) % 120 == 0:
                    self.run_autonomous_management_cycle()
                
                # Sleep until next heartbeat
                time.sleep(heartbeat_interval)
                
        except KeyboardInterrupt:
            logger.info(f"🛑 {self.agent_id} shutting down gracefully")
            self.send_message("system", f"{self.agent_id} is shutting down. Final status: {self.get_status_report()}")
        except Exception as e:
            logger.error(f"❌ {self.agent_id} error: {e}")
            self.send_message("system", f"{self.agent_id} encountered an error: {str(e)}")

def main():
    """Main entry point"""
    agent = AIContextManagerAgent()
    agent.run()

if __name__ == "__main__":
    main()
