pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-manager"
version = "1.0.0"
description = "Intelligent AI management system for any project"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "reshdesu", email = "reshdesu@example.com"},
]
maintainers = [
    {name = "reshdesu", email = "reshdesu@example.com"},
]
keywords = ["ai", "context", "management", "assistant", "development"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Tools",
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "anthropic>=0.67.0",
    "websockets>=15.0.1",
    "nltk>=3.9.1",
    "spacy>=3.8.7",
    "transformers>=4.56.1",
    "torch>=2.8.0",
    "aiohttp>=3.12.15",
    "psycopg2-binary>=2.9.9",
    "keyring>=24.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.0.280",
    "isort>=5.12",
    "mypy>=1.5",
    "flake8>=6.0",
    "bandit>=1.7",
    "pre-commit>=3.0",
    "types-PyYAML",
    "types-click",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=1.3",
]

[project.urls]
Homepage = "https://github.com/reshdesu/ai-manager"
Documentation = "https://github.com/reshdesu/ai-manager/wiki"
Repository = "https://github.com/reshdesu/ai-manager.git"
Issues = "https://github.com/reshdesu/ai-manager/issues"

[project.scripts]
ai-manager = "ai_manager.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_manager*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG", "S101"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ai_manager"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["ai_manager"]
omit = ["*/tests/*", "*/venv/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.18.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
    "types-click>=7.1.8",
    "types-pyyaml>=6.0.12.20250822",
]
